refactor to work with weird chrome web request behavior
Sometimes the web request completes before the tab state has started transitioning, which causes the icon/popup to reset when the tab finally transitions.

This new strategy waits for the web request and the tab before changing the icon, and using a single popup script that gets the current state from chrome.storage.local.